test/doca: split scrape helper into parser, model, and scrape loop
The scrape utility had grown a heavy std::regex parser inside the
timeSeries class. Separate the concerns into three cohesive headers:

  - timeseries.h: the time-series data model (labelSet, sample,
    seriesId, samples, seriesMap) plus the timeSeries query view. The
    view is now constructed from an already-parsed seriesMap, so it no
    longer knows how text becomes series -- decoupling it from parsing.
  - open_metrics_text_parser.h: OpenMetricsTextParser, turning a
    Prometheus / OpenMetrics text body into a seriesMap.
  - scrape_util.h: just scrapeUntilValue, the composition root that
    wires HTTP -> parse -> view.

The parser also tightens label-block validation: pairs must be
comma-separated and a trailing comma is rejected, so an exporter-format
regression surfaces as a missing series rather than being absorbed.
Addresses the remaining CodeRabbit review nit.

No test changes (the tests only use scrapeUntilValue + latestValue);
all DOCA telemetry tests pass.
